chore: enhasment in clear yarn command

+++ _enabled-commands/clear-yarn.sh
@@ -7,6 +7,7 @@
 
 # Optional parameters:
 # @raycast.icon 🤖
+# @raycast.argument1 { "type": "text", "placeholder": "clear node modules?" }
 # @raycast.packageName frontend-universe
 # @raycast.needsConfirmation true
 
@@ -15,12 +16,19 @@
 # @raycast.author Nirbhay Ram
 # @raycast.authorURL https://github.com/nirbhayram
 
-echo "removing parent node modules"
-cd ~/WebstormProjects/frontend-universe/
-rm -rf node_modules/
-cd ~/WebstormProjects/frontend-universe/packages/apps/classroom
-echo "removing classroom node modules"
-rm -rf node_modules/
-cd ~/WebstormProjects/frontend-universe/
+cd ~/WebstormProjects/frontend-universe/ || exit
+
+if [ "$1" = "yes" ]; then
+  echo "removing node modules inside frontend-universe directory"
+  rm -rf node_modules/
+fi
+
+cd ~/WebstormProjects/frontend-universe/packages/apps/classroom || exit
+if [ "$1" = "yes" ]; then
+  echo "removing node modules inside classroom directory"
+  rm -rf node_modules/
+fi
+
+cd ~/WebstormProjects/frontend-universe/ || exit
 echo "reinstall"
 yarn
